+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Custom sorting of reactions
 - Added a configurable separator view for new messages
 - Possibility to customize the cornerRadius of the `ComposerInputView`
+- Possibility to turn off tab bar handling in the message list
 
 ### 🐞 Fixed
 - Message List layout for iPad in Slide Over mode